requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
exifread
pysolar
requests
numpy
scipy
pydantic
